Move button out of settings and into an 'X' button on Homepage
