+++ app/src/main/res/layout/fragment_home.xml
@@ -396,6 +396,16 @@
                             android:text="@string/continue_watching" />
 
                     <ImageView
+                        app:tint="?attr/textColor"
+                        android:id="@+id/home_clear_watching"
+                        android:layout_marginEnd="40dp"
+                        android:layout_gravity="end|center_vertical"
+                        android:src="@drawable/ic_baseline_clear_24"
+                        android:layout_width="30dp"
+                        android:layout_height="match_parent"
+                        android:contentDescription="@string/clear_watching_pref" />
+
+                    <ImageView
                             app:tint="?attr/textColor"
                             android:layout_marginEnd="5dp"
                             android:layout_gravity="end|center_vertical"
